Consolidate adding logic
