+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentMethodType.swift
@@ -229,27 +229,22 @@ extension PaymentSheet {
             return paymentTypes
         }
 
+        /// Returns whether or not PaymentSheet, with the given `PaymentMethodRequirementProvider`s, should make the given `paymentMethod` available to add.
+        /// Note: This doesn't affect the availability of saved PMs.
+        /// - Parameters:
+        ///   - paymentMethod: the `STPPaymentMethodType` in question
+        ///   - requirementProviders: a list of [PaymentMethodRequirementProvider] who satisfy payment requirements
+        ///   - intent: a payment or setup intent
+        ///   - supportedPaymentMethods: the payment methods that PaymentSheet can display UI for
+        /// - Returns: true if `paymentMethod` should be available in the PaymentSheet, false otherwise
         static func supportsAdding(
             paymentMethod: PaymentMethodType,
             configuration: PaymentSheet.Configuration,
             intent: Intent,
             supportedPaymentMethods: [STPPaymentMethodType] = PaymentSheet.supportedPaymentMethods
         ) -> Bool {
-            switch intent {
-            case .paymentIntent(let paymentIntent):
-                // if using SFU on the PI fallthrough to the setup case
-                if paymentIntent.setupFutureUsage == .offSession || paymentIntent.setupFutureUsage == .onSession {
-                    fallthrough
-                }
-
-                if let stpPaymentMethodType = paymentMethod.stpPaymentMethodType {
-                    return PaymentSheet.supportsAdding(
-                        paymentMethod: stpPaymentMethodType,
-                        configuration: configuration,
-                        intent: intent,
-                        supportedPaymentMethods: supportedPaymentMethods
-                    )
-                } else if case .dynamic = paymentMethod {
+            guard let stpPaymentMethodType = paymentMethod.stpPaymentMethodType else {
+                if case .dynamic = paymentMethod {
                     return supports(
                         requirements: paymentMethod.supportsAddingRequirements(),
                         configuration: configuration,
@@ -257,41 +252,116 @@ extension PaymentSheet {
                     )
                 }
 
+                return false
+            }
+
+            switch intent {
+            case .paymentIntent(let paymentIntent):
+                if paymentIntent.setupFutureUsage != .none {
+                    fallthrough
+                }
             case .setupIntent:
-                return PaymentSheet.PaymentMethodType.supportsSaveAndReuse(
+                return supportsSaveAndReuse(
                     paymentMethod: paymentMethod,
                     configuration: configuration,
-                    intent: intent,
-                    supportedPaymentMethods: configuration.linkPaymentMethodsOnly
-                        ? PaymentSheet.supportedLinkPaymentMethods : PaymentSheet.supportedPaymentMethods
+                    intent: intent
                 )
             }
+
+            let requirements: [PaymentMethodTypeRequirement] = {
+                switch stpPaymentMethodType {
+                case .blik, .card, .cardPresent, .UPI, .weChatPay:
+                    return []
+                case .alipay, .EPS, .FPX, .giropay, .grabPay, .netBanking, .payPal, .przelewy24, .klarna,
+                    .linkInstantDebit, .bancontact, .iDEAL, .cashApp:
+                    return [.returnURL]
+                case .USBankAccount:
+                    return [
+                        .userSupportsDelayedPaymentMethods, .financialConnectionsSDK, .validUSBankVerificationMethod,
+                    ]
+                case .OXXO, .boleto, .AUBECSDebit, .SEPADebit:
+                    return [.userSupportsDelayedPaymentMethods]
+                case .bacsDebit, .sofort:
+                    return [.returnURL, .userSupportsDelayedPaymentMethods]
+                case .afterpayClearpay, .affirm:
+                    return [.returnURL, .shippingAddress]
+                case .link, .unknown:
+                    return [.unavailable]
+                @unknown default:
+                    return [.unavailable]
+                }
+            }()
+
+            return supports(
+                paymentMethod: stpPaymentMethodType,
+                requirements: requirements,
+                configuration: configuration,
+                intent: intent,
+                supportedPaymentMethods: supportedPaymentMethods
+            )
             // TODO: We need a way to model this information in our common model
-            return false
         }
 
+        /// Returns whether or not PaymentSheet should make the given `paymentMethod` available to save for future use, set up, and reuse
+        /// i.e. available for a PaymentIntent with setupFutureUsage or SetupIntent or saved payment method
+        /// - Parameters:
+        ///   - paymentMethod: the `STPPaymentMethodType` in question
+        ///   - requirementProviders: a list of [PaymentMethodRequirementProvider] who satisfy payment requirements
+        ///   - supportedPaymentMethods: the payment methods that PaymentSheet can display UI for
+        /// - Returns: true if `paymentMethod` should be available in the PaymentSheet, false otherwise
         static func supportsSaveAndReuse(
             paymentMethod: PaymentMethodType,
             configuration: PaymentSheet.Configuration,
             intent: Intent,
             supportedPaymentMethods: [STPPaymentMethodType] = PaymentSheet.supportedPaymentMethods
         ) -> Bool {
-            if let stpPaymentMethodType = paymentMethod.stpPaymentMethodType {
-                return PaymentSheet.supportsSaveAndReuse(
-                    paymentMethod: stpPaymentMethodType,
-                    configuration: configuration,
-                    intent: intent,
-                    supportedPaymentMethods: supportedPaymentMethods
-                )
-            } else if case .dynamic = paymentMethod {
-                return supports(
-                    requirements: paymentMethod.supportsSaveAndReuseRequirements(),
-                    configuration: configuration,
-                    intent: intent
-                )
+            guard let stpPaymentMethodType = paymentMethod.stpPaymentMethodType else {
+                if case .dynamic = paymentMethod {
+                    return supports(
+                        requirements: paymentMethod.supportsSaveAndReuseRequirements(),
+                        configuration: configuration,
+                        intent: intent
+                    )
+                }
+
+                return false
             }
+
+            let requirements: [PaymentMethodTypeRequirement] = {
+                switch stpPaymentMethodType {
+                case .card:
+                    return []
+                case .alipay:
+                    return [.returnURL]
+                case .USBankAccount:
+                    return [.userSupportsDelayedPaymentMethods]
+                case .iDEAL, .bancontact, .sofort:
+                    // SEPA-family PMs are disallowed until we can reuse them for PI+sfu and SI.
+                    // n.b. While iDEAL and bancontact are themselves not delayed, they turn into SEPA upon save, which IS delayed.
+                    return [.returnURL, .userSupportsDelayedPaymentMethods, .unavailable]
+                case .SEPADebit:
+                    // SEPA-family PMs are disallowed until we can reuse them for PI+sfu and SI.
+                    return [.userSupportsDelayedPaymentMethods, .unavailable]
+                case .bacsDebit:
+                    return [.returnURL, .userSupportsDelayedPaymentMethods]
+                case .AUBECSDebit, .cardPresent, .blik, .weChatPay, .grabPay, .FPX, .giropay, .przelewy24, .EPS,
+                    .netBanking, .OXXO, .afterpayClearpay, .payPal, .UPI, .boleto, .klarna, .link, .linkInstantDebit,
+                    .affirm, .cashApp, .unknown:
+                    return [.unavailable]
+                @unknown default:
+                    return [.unavailable]
+                }
+            }()
+
+            return supports(
+                paymentMethod: stpPaymentMethodType,
+                requirements: requirements,
+                configuration: configuration,
+                intent: intent,
+                supportedPaymentMethods: supportedPaymentMethods
+            )
+
             // TODO: We need a way to model this information in our common model
-            return false
         }
 
         static func supports(
@@ -304,6 +374,40 @@ extension PaymentSheet {
                 return accumulator + element.fulfilledRequirements
             }
             let supports = Set(requirements).isSubset(of: fulfilledRequirements)
+            return supports
+        }
+
+        /// DRY helper method
+        static func supports(
+            paymentMethod: STPPaymentMethodType,
+            requirements: [PaymentMethodTypeRequirement],
+            configuration: PaymentSheet.Configuration,
+            intent: Intent,
+            supportedPaymentMethods: [STPPaymentMethodType]
+        ) -> Bool {
+            guard supportedPaymentMethods.contains(paymentMethod) else {
+                return false
+            }
+
+            // Hide a payment method type if we are in live mode and it is unactivated
+            if !configuration.apiClient.isTestmode && intent.unactivatedPaymentMethodTypes.contains(paymentMethod) {
+                return false
+            }
+
+            let fulfilledRequirements = [configuration, intent].reduce([]) {
+                (accumulator: [PaymentMethodTypeRequirement], element: PaymentMethodRequirementProvider) in
+                return accumulator + element.fulfilledRequirements
+            }
+
+            let supports = Set(requirements).isSubset(of: fulfilledRequirements)
+            if paymentMethod == .USBankAccount {
+                if !fulfilledRequirements.contains(.financialConnectionsSDK) {
+                    print(
+                        "[Stripe SDK] Warning: us_bank_account requires the StripeConnections SDK. See https://stripe.com/docs/payments/ach-debit/accept-a-payment?platform=ios"
+                    )
+                }
+            }
+
             return supports
         }
     }
