+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+API.swift
@@ -271,7 +271,8 @@ extension PaymentSheet {
                         // Filter out payment methods that the PI/SI or PaymentSheet doesn't support
                         let savedPaymentMethods = paymentMethods
                             .filter { intent.recommendedPaymentMethodTypes.contains($0.type) }
-                            .filter { PaymentSheet.supportsSaveAndReuse(paymentMethod: $0.type, configuration: configuration, intent: intent) }
+                            .filter { PaymentSheet.supportsSaveAndReuse(paymentMethod: $0.paymentSheetPaymentMethodType(),
+                                                                        configuration: configuration, intent: intent) }
                         warnUnactivatedIfNeeded(unactivatedPaymentMethodTypes: intent.unactivatedPaymentMethodTypes)
 
                         let linkAccountPromise = PaymentSheet.lookupLinkAccount(
